requirements.txt

viam-sdk==0.55.1

adafruit-circuitpython-mpu6050==1.3.4

typing-extensions
